--- app/services/ContextIO/api_conversion_script.rb
@@ -1,20 +1,0 @@
-require 'ostruct'
-# This TX script checks to make sure the api post is either a json or
-# a hash and if it is a json then it parses it into a hash
-class ApiConversionScript < TransactionScript
-  def run(inputs)
-
-#inputs will be the raw params from the api post
-    inputs = check_for_json(inputs) if inputs.is_a?(String)
-
-#either inputs is a hash, is converted into a hash or it cannot
-#be converted into a hash and this will return a string
-    unless inputs.is_a?(Hash)
-      return failure 'Api POST request formatted incorrectly'
-    end
-
-    return success alert: inputs
-  end
-end
-
-
