--- spec/services/retrieve_message_script_spec.rb
@@ -1,24 +1,0 @@
-require 'spec_helper'
-
-describe RetrieveMessageScript do
-  it_behaves_like('TransactionScripts')
-
-  before(:all) do
-    @inputs = OpenStruct.new( alert:{'message_data'=> {'message_id'=> 1}, 'account_id' => 0})
-  end
-
-  it 'fails if the message is not a ContextIO::Message' do
-    allow_any_instance_of(ContextIO).to receive(:accounts){[OpenStruct.new(messages: ['here is the message'])]}
-    test = described_class.run(@inputs)
-    expect(test.error).to eq('message content missing')
-  end
-
-  xit 'fails if the message body is not a string' do
-    allow_any_instance_of(ContextIO).to receive(:accounts){[OpenStruct.new(messages: ['here is the message'])]}
-    allow(is_?).to
-    test = described_class.run(@inputs)
-    expect(test.error).to eq('message not formatted correctly')
-  end
-
-end
-
